requirements.txt
numpy==1.23.3
pandas==1.5.0
azure-kusto-data==3.1.3
matplotlib==3.6.0